# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeasurementComparisonResponse": {
        "properties": {
          "direction": {
            "$ref": "#/components/schemas/TraceDirection"
          },
          "excess_loss_db": {
            "description": "Perda excedente (previsto - medido). Valor positivo indica atenuação além da documentada em dB",
            "title": "Excess Loss Db",
            "type": "number"
          },
          "is_within_tolerance": {
            "description": "Indica se a medição de campo está dentro da margem de aceitação da engenharia",
            "title": "Is Within Tolerance",
            "type": "boolean"
          },
          "measured_power_dbm": {
            "description": "Potência real obtida pelo instrumento em dBm",
            "title": "Measured Power Dbm",
            "type": "number"
          },
          "measurement_id": {
            "title": "Measurement Id",
            "type": "string"
          },
          "predicted_power_dbm": {
            "description": "Potência teórica prevista pelo orçamento óptico em dBm",
            "title": "Predicted Power Dbm",
            "type": "number"
          },
          "tolerance_db": {
            "default": 2.0,
            "description": "Tolerância de desvio configurada em dB",
            "title": "Tolerance Db",
            "type": "number"
          },
          "wavelength_nm": {
            "title": "Wavelength Nm",
            "type": "integer"
          }
        },
        "required": [
          "measurement_id",
          "measured_power_dbm",
          "predicted_power_dbm",
          "excess_loss_db",
          "wavelength_nm",
          "direction",
          "is_within_tolerance"
        ],
        "title": "MeasurementComparisonResponse",
        "type": "object"
      },
      "MeasurementCreate": {
        "properties": {
          "direction": {
            "$ref": "#/components/schemas/TraceDirection",
            "default": "downstream",
            "description": "Direção do sinal medido: downstream (chegando na ONU) ou upstream (chegando na OLT)"
          },
          "instrument_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Fabricante e modelo do medidor óptico (Power Meter / OTDR)",
            "title": "Instrument Model"
          },
          "measured_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data e hora da coleta em campo (UTC)",
            "title": "Measured At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "power_dbm": {
            "description": "Potência óptica medida em dBm (ex: -21.50 dBm)",
            "title": "Power Dbm",
            "type": "number"
          },
          "service_link_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UUID do atendimento associado, se a medição ocorreu na ponta de um cliente",
            "title": "Service Link Id"
          },
          "terminal_id": {
            "description": "UUID do terminal óptico onde a medição foi realizada",
            "title": "Terminal Id",
            "type": "string"
          },
          "wavelength_nm": {
            "description": "Comprimento de onda calibrado no medidor (nm)",
            "maximum": 2000.0,
            "minimum": 800.0,
            "title": "Wavelength Nm",
            "type": "integer"
          }
        },
        "required": [
          "terminal_id",
          "power_dbm",
          "wavelength_nm"
        ],
        "title": "MeasurementCreate",
        "type": "object"
      },
      "MeasurementRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "direction": {
            "$ref": "#/components/schemas/TraceDirection"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "instrument_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument Model"
          },
          "measured_at": {
            "format": "date-time",
            "title": "Measured At",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "power_dbm": {
            "title": "Power Dbm",
            "type": "number"
          },
          "service_link_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Link Id"
          },
          "terminal_id": {
            "title": "Terminal Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "wavelength_nm": {
            "title": "Wavelength Nm",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "terminal_id",
          "power_dbm",
          "wavelength_nm",
          "direction",
          "measured_at",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "MeasurementRead",
        "type": "object"
      },
      "MeasurementUpdate": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "MeasurementUpdate",
        "type": "object"
      },
...
      "OpticalProfileCreate": {
        "properties": {
          "default_attenuation_db_per_km": {
            "default": 0.35,
            "description": "Atenuação nominal da fibra em dB/km para este comprimento de onda",
            "minimum": 0.0,
            "title": "Default Attenuation Db Per Km",
            "type": "number"
          },
          "name": {
            "description": "Nome do perfil óptico (ex: GPON Classe B+)",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "rx_overload_dbm": {
            "description": "Ponto de saturação/sobrecarga RX (dBm)",
            "title": "Rx Overload Dbm",
            "type": "number"
          },
          "rx_sensitivity_dbm": {
            "description": "Sensibilidade de recepção RX (dBm)",
            "title": "Rx Sensitivity Dbm",
            "type": "number"
          },
          "technology": {
            "default": "GPON",
            "description": "Tecnologia PON (GPON, XGS-PON, etc.)",
            "title": "Technology",
            "type": "string"
          },
          "tx_max_dbm": {
            "description": "Potência máxima de transmissão (dBm)",
            "title": "Tx Max Dbm",
            "type": "number"
          },
          "tx_min_dbm": {
            "description": "Potência mínima de transmissão (dBm)",
            "title": "Tx Min Dbm",
            "type": "number"
          },
          "wavelength_nm": {
            "description": "Comprimento de onda em nanômetros (ex: 1490)",
            "maximum": 2000.0,
            "minimum": 800.0,
            "title": "Wavelength Nm",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "wavelength_nm",
          "tx_min_dbm",
          "tx_max_dbm",
          "rx_sensitivity_dbm",
          "rx_overload_dbm"
        ],
        "title": "OpticalProfileCreate",
        "type": "object"
      },
...
      "OpticalSimulationRequest": {
        "properties": {
          "direction": {
            "$ref": "#/components/schemas/TraceDirection",
            "default": "downstream"
          },
          "engineering_margin_db": {
            "default": 3.0,
            "title": "Engineering Margin Db",
            "type": "number"
          },
          "overrides": {
            "description": "Lista de substituições pontuais a simular no caminho óptico",
            "items": {
              "$ref": "#/components/schemas/SimulationOverrideItem"
            },
            "minItems": 1,
            "title": "Overrides",
            "type": "array"
          },
          "service_link_id": {
            "title": "Service Link Id",
            "type": "string"
          }
        },
        "required": [
          "service_link_id",
          "overrides"
        ],
        "title": "OpticalSimulationRequest",
        "type": "object"
      },
...
      "PaginatedResponse_MeasurementRead_": {
        "properties": {
          "items": {
            "description": "Lista de itens da página atual",
            "items": {
              "$ref": "#/components/schemas/MeasurementRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "Página retornada",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Tamanho de página aplicado",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total global de registros correspondentes",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginatedResponse[MeasurementRead]",
        "type": "object"
      },
      "PaginatedResponse_OpticalProfileRead_": {
        "properties": {
          "items": {
            "description": "Lista de itens da página atual",
            "items": {
              "$ref": "#/components/schemas/OpticalProfileRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "Página retornada",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "description": "Tamanho de página aplicado",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total global de registros correspondentes",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginatedResponse[OpticalProfileRead]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "FTTH Manager — Sistema de documentação física e óptica de rede FTTH",
    "title": "FTTH Manager",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/measurements": {
      "get": {
        "operationId": "api_v1_measurements_list_measurements",
        "parameters": [
          {
            "in": "query",
            "name": "service_link_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Link Id"
            }
          },
          {
            "in": "query",
            "name": "terminal_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Terminal Id"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_MeasurementRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar medições de potência",
        "tags": [
          "Medições de Potência"
        ]
      },
      "post": {
        "operationId": "api_v1_measurements_create_measurement",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasurementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Registrar medição manual",
        "tags": [
          "Medições de Potência"
        ]
      }
    },
    "/api/v1/measurements/{measurement_id}": {
      "delete": {
        "operationId": "api_v1_measurements_measurement_id_delete_measurement",
        "parameters": [
          {
            "in": "path",
            "name": "measurement_id",
            "required": true,
            "schema": {
              "title": "Measurement Id",
              "type": "string"
            }
          },
          {
            "description": "Versão atual do recurso (If-Match)",
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "description": "Versão atual do recurso (If-Match)",
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Excluir medição",
        "tags": [
          "Medições de Potência"
        ]
      },
      "get": {
        "operationId": "api_v1_measurements_measurement_id_get_measurement",
        "parameters": [
          {
            "in": "path",
            "name": "measurement_id",
            "required": true,
            "schema": {
              "title": "Measurement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detalhes da medição",
        "tags": [
          "Medições de Potência"
        ]
      },
      "patch": {
        "operationId": "api_v1_measurements_measurement_id_update_measurement",
        "parameters": [
          {
            "in": "path",
            "name": "measurement_id",
            "required": true,
            "schema": {
              "title": "Measurement Id",
              "type": "string"
            }
          },
          {
            "description": "Versão atual do recurso (If-Match)",
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "description": "Versão atual do recurso (If-Match)",
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasurementUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualizar notas da medição",
        "tags": [
          "Medições de Potência"
        ]
      }
    },
    "/api/v1/measurements/{measurement_id}/compare": {
      "get": {
        "description": "Calcula a perda excedente em dB comparando o valor medido com o orçamento óptico da revisão correspondente.",
        "operationId": "api_v1_measurements_measurement_id_compare_compare_measurement",
        "parameters": [
          {
            "in": "path",
            "name": "measurement_id",
            "required": true,
            "schema": {
              "title": "Measurement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Comparar medição de campo com potência prevista",
        "tags": [
          "Medições de Potência"
        ]
      }
    },
    "/api/v1/optical-profiles": {
      "get": {
        "operationId": "api_v1_optical_profiles_list_optical_profiles",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_OpticalProfileRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar perfis ópticos",
        "tags": [
          "Cálculo Óptico e Perfis"
        ]
      },
      "post": {
        "operationId": "api_v1_optical_profiles_create_optical_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpticalProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpticalProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Criar perfil óptico",
        "tags": [
          "Cálculo Óptico e Perfis"
        ]
      }
    },
    "/api/v1/optical-profiles/{profile_id}": {
      "delete": {
        "operationId": "api_v1_optical_profiles_profile_id_delete_optical_profile",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "description": "Versão atual do recurso (If-Match)",
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "description": "Versão atual do recurso (If-Match)",
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Desativar perfil óptico",
        "tags": [
          "Cálculo Óptico e Perfis"
        ]
      },
      "get": {
        "operationId": "api_v1_optical_profiles_profile_id_get_optical_profile",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpticalProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detalhes do perfil óptico",
        "tags": [
          "Cálculo Óptico e Perfis"
        ]
      },
      "patch": {
        "operationId": "api_v1_optical_profiles_profile_id_update_optical_profile",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "description": "Versão atual do recurso (If-Match)",
            "in": "header",
            "name": "if-match",
            "required": true,
            "schema": {
              "description": "Versão atual do recurso (If-Match)",
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpticalProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpticalProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Atualizar perfil óptico",
        "tags": [
          "Cálculo Óptico e Perfis"
        ]
      }
    },
    "/api/v1/optical/budgets": {
      "post": {
        "description": "Calcula atenuação acumulada, potência RX prevista, margem de engenharia e sobrecarga para um atendimento documentado.",
        "operationId": "api_v1_optical_budgets_calculate_budget",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetCalculationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetCalculationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calcular orçamento de potência óptica (Link Budget)",
        "tags": [
          "Cálculo Óptico e Perfis"
        ]
      }
    },
    "/api/v1/optical/simulations": {
      "post": {
        "description": "Compara o orçamento de potência original com um cenário hipotético que aplica overrides em perdas, comprimentos ou splitters.",
        "operationId": "api_v1_optical_simulations_simulate_budget",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpticalSimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpticalSimulationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulação óptica com parâmetros hipotéticos",
        "tags": [
          "Cálculo Óptico e Perfis"
        ]
      }
    },
...
  }
}